requirements.txt
aiogram>=3.0.0,<4.0.0
python-dotenv